LayoutTests/fast/shadow-dom/mutation-event-in-shadow-tree-expected.txt

This tests mutating shadow tree inside execCommand. WebKit should not dispach DOM mutation events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mutationEventsWith((element) => element.setAttribute('class', 'foo')) is "[]"
PASS mutationEventsWith((element) => element.appendChild(document.createElement('span'))) is "[]"
PASS mutationEventsWith((element) => { element.textContent = 'hello'; element.firstChild.appendData(' world') }) is "[]"
PASS mutationEventsWith((element) => element.appendChild(element.appendChild(document.createElement('span')))) is "[]"
PASS successfullyParsed is true

TEST COMPLETE
